modify code

+++ src/class049/Code02_DiffMatrixLuogu.java
@@ -4,7 +4,7 @@ package class049;
 // 测试链接 : https://www.luogu.com.cn/problem/P3397
 // 请同学们务必参考如下代码中关于输入、输出的处理
 // 这是输入输出处理效率很高的写法
-// 提交以下的code，提交时请把类名改成"Main"
+// 提交以下的code，提交时请把类名改成"Main"，可以直接通过
 
 import java.io.BufferedReader;
 import java.io.IOException;
@@ -13,12 +13,11 @@ import java.io.OutputStreamWriter;
 import java.io.PrintWriter;
 import java.io.StreamTokenizer;
 
-// 洛谷对java语言不友好，经常卡空间，多提交几次能全部通过的
 public class Code02_DiffMatrixLuogu {
 
 	public static int MAXN = 1002;
 
-	public static long[][] diff = new long[MAXN][MAXN];
+	public static int[][] diff = new int[MAXN][MAXN];
 
 	public static int n, q;
